+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/X509Certificates/OpenSslX509ChainProcessor.cs
@@ -12,7 +12,6 @@ using System.Security.Cryptography.X509Certificates.Asn1;
 using System.Text;
 using Microsoft.Win32.SafeHandles;
 using Internal.Cryptography;
-using Interop.Crypto;
 
 using X509VerifyStatusCodeUniversal = Interop.Crypto.X509VerifyStatusCodeUniversal;
 
@@ -1030,7 +1029,7 @@ namespace System.Security.Cryptography.X509Certificates
             List<X509ChainStatus> overallStatus,
             ref bool overallHasNotSignatureValid)
         {
-            foreach (X509VerifyStatusCode errorCode in errorCodes)
+            foreach (Interop.Crypto.X509VerifyStatusCode errorCode in errorCodes)
             {
                 AddElementStatus(errorCode, elementStatus, overallStatus, ref overallHasNotSignatureValid);
             }
@@ -1085,6 +1084,7 @@ namespace System.Security.Cryptography.X509Certificates
 
             elementStatus.Add(chainStatus);
             AddUniqueStatus(overallStatus, ref chainStatus);
+
             if (statusFlag == X509ChainStatusFlags.NotSignatureValid)
             {
                 overallHasNotSignatureValid = true;
